+++ src/Exception/AccessException.php
@@ -19,6 +19,6 @@
 
 namespace Art4\JsonApiClient\Exception;
 
-class AccessException extends \RuntimeException implements Exception
+class AccessException extends \RuntimeException implements Exception, \Throwable
 {
 }