+++ src/Exception/Exception.php
@@ -19,6 +19,10 @@
 
 namespace Art4\JsonApiClient\Exception;
 
+/**
+ * @deprecated `\Art4\JsonApiClient\Exception\Exception` will extend `\Throwable` in v2.0. Do the same in your implementation now to avoid errors.
+ */
+// interface Exception extends \Throwable
 interface Exception
 {
 }