+++ src/Factory.php
@@ -19,8 +19,6 @@
 
 namespace Art4\JsonApiClient;
 
-use Art4\JsonApiClient\Accessable;
-
 /**
  * Factory Interface
  */
@@ -29,10 +27,14 @@ interface Factory
     /**
      * Create a new instance of a class
      *
+     * @deprecated The `\Art4\JsonApiClient\Factory::make()` methods first parameter signature will be `string` in v2.0.
+     * @deprecated `\Art4\JsonApiClient\Factory::make()` will add `\Art4\JsonApiClient\Accessable` as a native return type declaration in v2.0. Do the same in your implementation now to avoid errors.
+     *
      * @param string $name
      * @param array<mixed|Manager|Accessable>  $args
      *
      * @return \Art4\JsonApiClient\Accessable
      */
     public function make($name, array $args = []);
+    // public function make(string $name, array $args = []): Accessable;
 }